requirements.txt
bootstrap4==0.1.0
click==7.1.2
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Moment==0.11.0
Flask-Script==2.0.6
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
